// types/Product.ts
export interface Product {
    id: string;
    title: string;
    price: number;
    zipcode: string;
    seller: string;
    thumbnailHd: string;
    date: string;
    description: string;
    category: string;
    stock: number;
    rating: number;
    reviewsCount: number;
    brand: string;
    sku: string;
    weight: number;
    dimensions: string;
    images: string[];
    highlight: boolean;
}